Split compiler tests
